refactore

+++ Application/Rooms/RoomCommands/RoomCommandHandlers/DeleteRoomHandler.cs
@@ -1,3 +1,4 @@
+using Application.Result;
 using Application.Rooms.RoomCommands.RoomCommandRequests;
 using Application.Rooms.RoomDtos;
 using AutoMapper;
@@ -6,14 +7,17 @@ using MediatR;
 
 namespace Application.Rooms.RoomCommands.RoomCommandHandlers;
 
-public class DeleteRoomHandler(IRoomRepository roomRepository, IMapper mapper) : IRequestHandler<DeleteRoomCommand, RoomDto?>
+public class DeleteRoomHandler(IRoomRepository roomRepository, IMapper mapper)
+    : IRequestHandler<DeleteRoomCommand, Result<RoomDto>>
 {
-    public async Task<RoomDto?> Handle(DeleteRoomCommand request, CancellationToken cancellationToken)
+    public async Task<Result<RoomDto>> Handle(DeleteRoomCommand request, CancellationToken cancellationToken)
     {
-        var room = await roomRepository.DeleteAsync(request.RoomId, cancellationToken);
-        if(room == null)
-            return null;
-        var roomDto = mapper.Map<RoomDto>(room);
-        return  roomDto;
+        if (await roomRepository.GetByIdAsync(request.RoomId, cancellationToken) == null)
+            return Result<RoomDto>.Failure($"room {request.RoomId} not found", 404);
+        var result = await roomRepository.DeleteAsync(request.RoomId, cancellationToken);
+        if (result == null)
+            return Result<RoomDto>.Failure($"insert room failed", 400);
+        var roomDto = mapper.Map<RoomDto>(result);
+        return Result<RoomDto>.Success(roomDto);
     }
 }
